+++ src/main/kotlin/site/siredvin/turtlematic/util/GlobalFlags.kt
@@ -1,0 +1,5 @@
+package site.siredvin.turtlematic.util
+
+object GlobalFlags {
+    val RAIL_SHAPE_CORRECTION_SUPPRESSION: ThreadLocal<Boolean> = ThreadLocal()
+}